+++ test/runtests.jl
@@ -4,7 +4,6 @@ using JlrsCore.Ledger
 using Test
 
 include("Ledger.jl")
-include("RustResult.jl")
 include("AbstractTypes.jl")
 include("SingleFieldBits.jl")
 include("Mutable.jl")