+++ CHANGELOG.md
@@ -1,3 +1,17 @@
+#### v0.4.0
+
+- `RustResult` has been removed.
+
+- Support for async callbacks and their thread pools have been removed.
+
+- `Reflect.reflect` skips generating bindings for mutable types that are only used as field types, i.e. `Reflect.reflect([T])` now only generates bindings that are necessary to represent the layout of `T`.
+
+- The `internaltypes` keyword argument has been removed from `Reflect.reflect` because jlrs has dropped support for these types. `Expr` has been promoted to a "regular" managed type, other internal types now map to `Option<ValueRef>`.
+
+- `Wrap.@wrapmodule` supports generating functions that are generic over one or more parameters, and restricting those parameters.
+
+- `BackgroundTask` and `DelegatedTask` replace the old async callbacks.
+
 #### v0.3.0
 
 - The `IsBits` trait is derived for layout types by `Reflect.reflect` if the type it reflects is an `isbits` type when all type parameters that affect the layout are `isbits` types.